Store and dashboard cleaned up. No duplicate items allowed anymore. Delete added to debts. Invoiced items added to debts where invoices exist. Store now served out of invoice controller. Expensing out of store page now creates a debt with the cost provided and auto-splits with all users.

+++ app/controllers/debts_controller.rb
@@ -42,8 +42,15 @@ class DebtsController < ApplicationController
 
   def destroy
     @debt = Debt.find(params[:id])
+    if current_user.id != @debt.owner.id
+      redirect_to dashboard_sessions_path
+      return
+    end
     @debt.debtors.each do |debtor|
       debtor.destroy
+    end
+    if @debt.invoice
+      @debt.invoice.destroy
     end
     @debt.destroy
     redirect_to dashboard_sessions_path
